Revert "Move file distribution stuff to filedistribution module"
