+++ config-proxy/src/main/java/com/yahoo/vespa/config/proxy/ProxyServer.java
@@ -5,8 +5,6 @@ import com.yahoo.concurrent.DaemonThreadFactory;
 import com.yahoo.config.subscription.ConfigSourceSet;
 import com.yahoo.jrt.Spec;
 
-import com.yahoo.jrt.Supervisor;
-import com.yahoo.jrt.Transport;
 import com.yahoo.log.LogLevel;
 import com.yahoo.log.LogSetup;
 import com.yahoo.log.event.Event;
@@ -42,7 +40,6 @@ public class ProxyServer implements Runnable {
 
     // Scheduled executor that periodically checks for requests that have timed out and response should be returned to clients
     private final ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1, new DaemonThreadFactory());
-    private final Supervisor supervisor = new Supervisor(new Transport());
     private final ClientUpdater clientUpdater;
     private ScheduledFuture<?> delayedResponseScheduler;
 
@@ -87,7 +84,6 @@ public class ProxyServer implements Runnable {
         clientUpdater = new ClientUpdater(rpcServer, statistics, delayedResponses);
         this.configClient = createClient(clientUpdater, delayedResponses, source, timingValues, memoryCache, configClient);
         this.fileDownloader = new FileDownloader(new JRTConnectionPool(source));
-        new com.yahoo.vespa.filedistribution.RpcServer(supervisor, fileDownloader);
     }
 
     static ProxyServer createTestServer(ConfigSourceSet source) {
@@ -166,7 +162,7 @@ public class ProxyServer implements Runnable {
     }
 
     private ConfigProxyRpcServer createRpcServer(Spec spec) {
-        return  (spec == null) ? null : new ConfigProxyRpcServer(this, supervisor, spec); // TODO: Try to avoid first argument being 'this'
+        return  (spec == null) ? null : new ConfigProxyRpcServer(this, spec); // TODO: Try to avoid first argument being 'this'
     }
 
     private RpcConfigSourceClient createRpcClient() {
