+++ config-proxy/src/main/java/com/yahoo/vespa/config/proxy/filedistribution/FileDownloader.java
