+++ config-proxy/src/test/java/com/yahoo/vespa/config/proxy/ConfigProxyRpcServerTest.java
@@ -5,8 +5,6 @@ import com.yahoo.config.subscription.ConfigSourceSet;
 import com.yahoo.jrt.Request;
 import com.yahoo.jrt.Spec;
 import com.yahoo.jrt.StringValue;
-import com.yahoo.jrt.Supervisor;
-import com.yahoo.jrt.Transport;
 import com.yahoo.vespa.config.RawConfig;
 import org.junit.After;
 import org.junit.Before;
@@ -30,7 +28,7 @@ public class ConfigProxyRpcServerTest {
     @Before
     public void setup() {
         proxyServer = ProxyServer.createTestServer(new ConfigSourceSet(address));
-        rpcServer = new ConfigProxyRpcServer(proxyServer, new Supervisor(new Transport()), null);
+        rpcServer = new ConfigProxyRpcServer(proxyServer, null);
     }
 
     @After
@@ -42,7 +40,7 @@ public class ConfigProxyRpcServerTest {
     public void basic() {
         ProxyServer proxy = ProxyServer.createTestServer(new MockConfigSource(new MockClientUpdater()));
         Spec spec = new Spec("localhost", 12345);
-        ConfigProxyRpcServer server = new ConfigProxyRpcServer(proxy, new Supervisor(new Transport()), spec);
+        ConfigProxyRpcServer server = new ConfigProxyRpcServer(proxy, spec);
         assertThat(server.getSpec(), is(spec));
     }
