+++ configserver/pom.xml
@@ -126,11 +126,6 @@
     </dependency>
     <dependency>
       <groupId>com.yahoo.vespa</groupId>
-      <artifactId>filedistribution</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.yahoo.vespa</groupId>
       <artifactId>filedistributionmanager</artifactId>
       <version>${project.version}</version>
     </dependency>
